requirements.txt
asgiref==3.11.0
detect==2020.12.3
Django==6.0.1
django-detect==1.0.20
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
django-six==1.0.5
djangorestframework==3.16.1
sqlparse==0.5.5
django-environ
